requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
numpy>=1.23.2,<2.0.0
setuptools==70.2.0
river==0.21.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
xgboost==2.1.1